added hook useFocusedObject to handle current focused field
