+++ src/components/Forms/SizesForm/Steps/Step3/Window.tsx
@@ -1,6 +1,5 @@
 'use client';
-import { useEffect, useMemo, useState } from 'react';
-import { useAppDispatch } from '@/redux/hooks';
+import { useEffect, useMemo } from 'react';
 import { useFormContext } from 'react-hook-form';
 import ClearIcon from '@mui/icons-material/Clear';
 import { RadioGroupWrapper } from '@/components/Input/RadioGroup/RadioGroupWrapper';
@@ -12,10 +11,7 @@ import {
 } from '@/components/Input/classNameConstants';
 import { Button, FormHelperText, Stack, Typography } from '@mui/material';
 import { useToWallRadios } from '@/hooks/useToWallRadios';
-import {
-  addBedroomFocusedField,
-  deleteBedroomFocusedField,
-} from '@/redux/slices/focusedFields-slice';
+import { useFocusedObject } from '@/hooks/useFocusedObject';
 import {
   MAX_WALLS_INPUT_LENGTH,
   MAX_WINDOW_INPUT_LENGTH,
@@ -35,10 +31,9 @@ export const Window = ({ index, handleRemove }: WindowProps) => {
     watch,
     setValue,
   } = useFormContext<SizesFormType>();
-  const dispatch = useAppDispatch();
-  const [isHovered, setHovered] = useState<boolean>(false);
-  const [isFocused, setFocused] = useState<boolean>(false);
-
+  const { handleSetFocused, handleSetHovered } = useFocusedObject(
+    `windows.windows.${index}.${STEP3.windowSize.name}`,
+  );
   const selectedWall = watch(
     `windows.windows.${index}.${STEP3.wallNumber.name}`,
   );
@@ -55,18 +50,6 @@ export const Window = ({ index, handleRemove }: WindowProps) => {
         shouldValidate: true,
       });
   }, [selectedWall, setValue, index]);
-
-  useEffect(() => {
-    if (isHovered || isFocused) {
-      dispatch(
-        addBedroomFocusedField(
-          `windows.windows.${index}.${STEP3.windowSize.name}`,
-        ),
-      );
-    } else {
-      dispatch(deleteBedroomFocusedField());
-    }
-  }, [dispatch, isHovered, isFocused, index]);
 
   const options = useMemo(() => {
     const doorWallNumber = watch('door.wallNumber');
@@ -87,8 +70,8 @@ export const Window = ({ index, handleRemove }: WindowProps) => {
       </Stack>
       <Stack
         gap={3}
-        onMouseEnter={() => setHovered(true)}
-        onMouseLeave={() => setHovered(false)}
+        onMouseEnter={() => handleSetHovered(true)}
+        onMouseLeave={() => handleSetHovered(false)}
       >
         <SelectWrapper
           name={`windows.windows.${index}.${STEP3.wallNumber.name}`}
@@ -110,8 +93,8 @@ export const Window = ({ index, handleRemove }: WindowProps) => {
               errors?.windows?.windows?.[index]?.size?.message) ||
             ''
           }
-          onFocus={() => setFocused(true)}
-          onBlurHandler={() => setFocused(false)}
+          onFocus={() => handleSetFocused(true)}
+          onBlurHandler={() => handleSetFocused(false)}
         />
         <Stack>
           <Stack direction="row" gap={3}>
@@ -123,8 +106,8 @@ export const Window = ({ index, handleRemove }: WindowProps) => {
               type="number"
               step={1}
               max={MAX_WALLS_INPUT_LENGTH}
-              onFocus={() => setFocused(true)}
-              onBlurHandler={() => setFocused(false)}
+              onFocus={() => handleSetFocused(true)}
+              onBlurHandler={() => handleSetFocused(false)}
             />
             <RadioGroupWrapper
               name={`windows.windows.${index}.${STEP3.toWall.name}`}
