+++ src/components/Forms/SizesForm/Steps/Step3/WindowWithBalcony.tsx
@@ -1,6 +1,5 @@
 'use client';
-import { useEffect, useMemo, useState } from 'react';
-import { useAppDispatch } from '@/redux/hooks';
+import { useEffect, useMemo } from 'react';
 import { useFormContext } from 'react-hook-form';
 import ClearIcon from '@mui/icons-material/Clear';
 import { RadioGroupWrapper } from '@/components/Input/RadioGroup/RadioGroupWrapper';
@@ -12,10 +11,7 @@ import {
 } from '@/components/Input/classNameConstants';
 import { Button, FormHelperText, Stack, Typography } from '@mui/material';
 import { useToWallRadios } from '@/hooks/useToWallRadios';
-import {
-  addBedroomFocusedField,
-  deleteBedroomFocusedField,
-} from '@/redux/slices/focusedFields-slice';
+import { useFocusedObject } from '@/hooks/useFocusedObject';
 import {
   MAX_DOOR_INPUT_LENGTH,
   MAX_WALLS_INPUT_LENGTH,
@@ -39,10 +35,9 @@ export const WindowWithBalcony = ({
     watch,
     setValue,
   } = useFormContext<SizesFormType>();
-  const dispatch = useAppDispatch();
-  const [isHovered, setHovered] = useState<boolean>(false);
-  const [isFocused, setFocused] = useState<boolean>(false);
-
+  const { handleSetFocused, handleSetHovered } = useFocusedObject(
+    `windows.windows.${index}.${STEP3.windowSize.name}`,
+  );
   const selectedWall = watch(
     `windows.windows.${index}.${STEP3.wallNumber.name}`,
   );
@@ -59,18 +54,6 @@ export const WindowWithBalcony = ({
         shouldValidate: true,
       });
   }, [selectedWall, setValue, index]);
-
-  useEffect(() => {
-    if (isHovered || isFocused) {
-      dispatch(
-        addBedroomFocusedField(
-          `windows.windows.${index}.${STEP3.windowSize.name}`,
-        ),
-      );
-    } else {
-      dispatch(deleteBedroomFocusedField());
-    }
-  }, [dispatch, isHovered, isFocused, index]);
 
   const options = useMemo(() => {
     const doorWallNumber = watch('door.wallNumber');
@@ -100,8 +83,8 @@ export const WindowWithBalcony = ({
         <Stack
           gap={3}
           width="100%"
-          onMouseEnter={() => setHovered(true)}
-          onMouseLeave={() => setHovered(false)}
+          onMouseEnter={() => handleSetHovered(true)}
+          onMouseLeave={() => handleSetHovered(false)}
         >
           <TextFieldWrapper
             name={`windows.windows.${index}.${STEP3.windowSize.name}`}
@@ -116,8 +99,8 @@ export const WindowWithBalcony = ({
                 errors?.windows?.windows?.[index]?.size?.message) ||
               ''
             }
-            onFocus={() => setFocused(true)}
-            onBlurHandler={() => setFocused(false)}
+            onFocus={() => handleSetFocused(true)}
+            onBlurHandler={() => handleSetFocused(false)}
           />
           <TextFieldWrapper
             name={`windows.windows.${index}.${STEP3.doorSize.name}`}
@@ -132,8 +115,8 @@ export const WindowWithBalcony = ({
                 errors?.windows?.windows?.[index]?.doorSize?.message) ||
               ''
             }
-            onFocus={() => setFocused(true)}
-            onBlurHandler={() => setFocused(false)}
+            onFocus={() => handleSetFocused(true)}
+            onBlurHandler={() => handleSetFocused(false)}
           />
 
           <Stack>
@@ -146,8 +129,8 @@ export const WindowWithBalcony = ({
                 type="number"
                 step={1}
                 max={MAX_WALLS_INPUT_LENGTH}
-                onFocus={() => setFocused(true)}
-                onBlurHandler={() => setFocused(false)}
+                onFocus={() => handleSetFocused(true)}
+                onBlurHandler={() => handleSetFocused(false)}
               />
               <RadioGroupWrapper
                 name={`windows.windows.${index}.${STEP3.toWall.name}`}
